gestao contribuintes

+++ src/main/java/br/com/gestao_contribuintes/gestaocontribuintes/Service/ContribuintesService.java
@@ -4,78 +4,71 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
-import jakarta.transaction.Transactional;
+//import jakarta.transaction.Transactional;
 
 import br.com.gestao_contribuintes.gestaocontribuintes.Entity.Contribuintes;
-import br.com.gestao_contribuintes.gestaocontribuintes.Entity.Filiacao;
-import br.com.gestao_contribuintes.gestaocontribuintes.Entity.Responsavel;
+import br.com.gestao_contribuintes.gestaocontribuintes.Entity.Dependentes;
+//import br.com.gestao_contribuintes.gestaocontribuintes.Entity.Filiacao;
+//import br.com.gestao_contribuintes.gestaocontribuintes.Entity.Responsavel;
 import br.com.gestao_contribuintes.gestaocontribuintes.Repository.ContribuintesRepository;
-import br.com.gestao_contribuintes.gestaocontribuintes.Repository.FiliacaoRepository;
-import br.com.gestao_contribuintes.gestaocontribuintes.Repository.ResponsavelRepository;
+//import br.com.gestao_contribuintes.gestaocontribuintes.Repository.FiliacaoRepository;
+//import br.com.gestao_contribuintes.gestaocontribuintes.Repository.ResponsavelRepository;
 
 @Service
 public class ContribuintesService {
 
     private final ContribuintesRepository contribuintesRepository;
-    private final FiliacaoRepository filiacaoRepository;
-    private ResponsavelRepository responsavelRepository;
 
-    public ContribuintesService(ContribuintesRepository contribuintesRepository, FiliacaoRepository filiacaoRepository,
-            ResponsavelRepository responsavelRepository) {
+    public ContribuintesService(ContribuintesRepository contribuintesRepository) {
         this.contribuintesRepository = contribuintesRepository;
-        this.filiacaoRepository = filiacaoRepository;
-        this.responsavelRepository = responsavelRepository;
     }
 
-    @SuppressWarnings("null")
     public Contribuintes create(Contribuintes contribuintes) {
         return contribuintesRepository.save(contribuintes);
     }
 
-    @Transactional
     public List<Contribuintes> getAllContribuintes() {
         return contribuintesRepository.findAll();
     }
 
     public Optional<Contribuintes> update(Contribuintes contribuintes) {
-        // Verifica se o contribuinte existe antes de atualizar
         if (contribuintes.getCPF() == null || !contribuintesRepository.existsByCPF(contribuintes.getCPF())) {
-            return Optional.empty(); // Retorna um Optional vazio se o contribuinte não existir
+            return Optional.empty();
         }
         return Optional.of(contribuintesRepository.save(contribuintes));
     }
 
-    @Transactional
     public boolean delete(String cpf) {
-        if (cpf != null) { // Verifica se o CPF não é nulo
-            if (contribuintesRepository.existsByCPF(cpf)) {
-                contribuintesRepository.deleteByCPF(cpf);
-                return true; // Retorna true se o contribuinte for excluído com sucesso
-            }
+        if (cpf != null && contribuintesRepository.existsByCPF(cpf)) {
+            contribuintesRepository.deleteByCPF(cpf);
+            return true;
         }
-        return false; // Retorna false se o CPF for nulo ou o contribuinte não existir
+        return false;
     }
 
-    public Filiacao cadastrarDependente(String cpfResponsavel, Filiacao dependente) {
-        Optional<Responsavel> responsavelOptional = responsavelRepository.findById(cpfResponsavel);
-        if (responsavelOptional.isEmpty()) {
-            throw new IllegalArgumentException("O responsável com o CPF " + cpfResponsavel + " não foi encontrado.");
+    public Contribuintes addDependente(String cpfContribuinte, Dependentes dependente) {
+        Optional<Contribuintes> contribuinteOptional = contribuintesRepository.findById(cpfContribuinte);
+        if (contribuinteOptional.isEmpty()) {
+            throw new IllegalArgumentException("O contribuinte com o CPF " + cpfContribuinte + " não foi encontrado.");
         }
 
-        Contribuintes responsavel = contribuintesRepository.findByCPF(cpfResponsavel);
-        dependente.setContribuinte(responsavel);
-        dependente.setCPF(cpfResponsavel);
+        Contribuintes contribuinte = contribuinteOptional.get();
+        List<Dependentes> dependentes = contribuinte.getDependentes();
+        dependentes.add(dependente);
+        contribuinte.setDependentes(dependentes);
 
-        return filiacaoRepository.save(dependente);
+        return contribuintesRepository.save(contribuinte);
     }
 
     public boolean existsByCPF(String cpf) {
         return contribuintesRepository.existsByCPF(cpf);
     }
 
-    public List<Filiacao> getDependentesByContribuinteCPF(String cpf) {
-        // Consulta o repositório para obter a lista de dependentes associados ao
-        // contribuinte com o CPF fornecido
-        return filiacaoRepository.findByContribuinte_CPF(cpf);
+    public List<Dependentes> getDependentesByContribuinteCPF(String cpf) {
+        Optional<Contribuintes> contribuinteOptional = contribuintesRepository.findById(cpf);
+        if (contribuinteOptional.isEmpty()) {
+            throw new IllegalArgumentException("O contribuinte com o CPF " + cpf + " não foi encontrado.");
+        }
+        return contribuinteOptional.get().getDependentes();
     }
 }
